rust/ffi: move helper trait state_get_tx_iterator to ffi
Ticket: 7666
